backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.36
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.13.1
python-dotenv==1.0.0
httpx==0.27.2
requests==2.31.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
textstat==0.7.3
python-slugify==8.0.4
transformers==4.43.3
aiosqlite==0.19.0
asyncpg==0.30.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==7.0.0
torch>=2.0.0; platform_system!="Windows" or platform_machine!="x86_64"
PyMuPDF==1.24.9
pdfplumber==0.11.0
sentence-transformers==2.2.2
numpy<2.0,>=1.17.3
duckduckgo-search==5.3.1
networkx==3.2.1
gensim>=4.3.3
# node2vec==0.4.6  # Optional: Install manually if needed for graph embeddings (requires numpy<2.0, incompatible with Python 3.13)
# Note: Using gensim Word2Vec with custom Node2Vec implementation for Python 3.13 compatibility

# Phase 6.5: Advanced Metadata Extraction Dependencies
camelot-py[base]==0.11.0
tabula-py==2.9.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.4.0
sympy==1.12
nltk==3.8.1
python-Levenshtein==0.25.0
scikit-learn==1.3.2

# Production ML Training Dependencies
arxiv>=1.4.0
optuna>=3.0.0
tensorboard>=2.14.0

# Phase 12.5: Event-Driven Architecture Dependencies
celery[redis]>=5.3.0
flower>=2.0.0
redis>=5.0.0

# Phase 12.6: Test Suite Optimization Dependencies
pytest-xdist>=3.5.0

# Phase 14.5: Property-Based Testing Dependencies
hypothesis>=6.0.0

# Phase 17: JWT Authentication Dependencies
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==4.1.3  # Pin to compatible version with passlib

# Phase 17.5: Advanced RAG Dependencies
spacy>=3.8.0
FlagEmbedding>=1.2.0

# Phase 19: Resilience Patterns
pybreaker>=1.0.0

# Phase 18: Code Intelligence Dependencies
tree-sitter==0.23.2
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
tree-sitter-typescript==0.23.2
tree-sitter-rust==0.23.2
tree-sitter-go==0.23.3
tree-sitter-java==0.23.4
gitpython==3.1.40
pathspec==0.12.1
